migrate-diff: error on non-existing sqlite database
The default behaviour of the sqlite driver is to implicitly create the
database when we connect to it.

Given that

- db push has a database creation step
- implicit database creation is otherwise unexpected
- there is a stub of this for some commands in the TS code already
- we have more than enough time to QA changes we make today

This commit changes the sql migration connector so that whenever we
connect to a sqlite database, we return a P1003 (DatabaseDoesNotExist)
error if there is no file at the expected path.

closes https://github.com/prisma/prisma/issues/12226