Implement cockroachdb autoincrement columns proposal
This PR implements the proposal on cockroachdb autoincrementing columns for
the introspection and migration engines, and adapts tests where
necessary in the query engine test suite.

This PR is a best effort implementation given the time constraints — it
makes a number of design and implementation compromises to ship the
changes.

**Overview of the changes**

On the cockroachdb connector:

- `@default(autoincrement())` now maps to the default mechanism in
  cockroachdb, `DEFAULT unique_rowid()`. Since that makes the column
  a 64 bit integer column automatically, this
  `@default(autoincrement())` is only supported on `BigInt` fields.`
- A new `@default(sequence())` maps to the usage of sequences in the
  database, which is not the recommended default on cockroachdb but
  closely matches how postgres autoincrement columns work. The
  `sequence()` function can take parameters defined in the PSL proposal.

**Tests**

The main difficulty is with the tests. Now that we map
`@default(autoincrement())` to `DEFAULT unique_rowid()` on cockroachdb
(implicitly through the use of SERIAL), autoincrementing integers in
cockroachdb prisma schemas cannot be `Int` anymore, only `BigInt`. This
makes most of the introspection and migration tests that are portable
between databases not portable to cockroachdb. In this PR, we pick a
subset we still want to test and adapt them to cockroachdb, and ignore
the rest. This is obviously less than ideal.

In the query engine tests, we adapt this similarly, with the additional
difficulty that values produced by unique_rowid() are not stable and do
not start at one. The tests that were not passing are special-cased for
cockroachdb and spread between schemas with `@default(sequence())`
(closely matches postgres @default(autoincrement())) and tests that use
the cockroach default `@default(unique_rowid())`.

An introspection and migrations test setup rewrite initiative is in the
plans.

**Limitations**

- In migrations, we do not diff the settings of existing sequences yet.

closes https://github.com/prisma/prisma/issues/12239
closes https://github.com/prisma/prisma/issues/12113
closes https://github.com/prisma/prisma/issues/12244