feat(qe) Metrics
Adds metrics to the query engine. The metrics are exposed via the
http server `/metrics` endpoint. For Metrics to be enabled the flag
`--enable-metrics` needs to be set and `"metrics"` added to the
`preview_features`.

Metrics can be outputted as `json` or `prometheus` format.
/metrics -> returns prometheus format
/metrics?format=json -> returns metrics in json format.

fix failing unit test

add tx metric test

clippy fixes