refactor(me): stop using async_trait in migration connectors
- It produces worse compiler errors
- It obscures what is actually going on
- It's an extra dependency that brings in the whole proc macro machinery

I measured single-crate compile times for sql-migration connector before
942313679488ec2183105c6b3b2aa9d09459254c, at
942313679488ec2183105c6b3b2aa9d09459254c and with this commit. Together,
the two commits represent a ~1 second improvement, with each one
contributing about half of it.

This commit also includes the removal of the `authors` fields in the
migration-engine crates, like we've started doing elsewhere (and as per
the new cargo defaults).