refactor(datamodel): make dml_to_ast ready for new param
In order to fix https://github.com/prisma/prisma/issues/12998, we want
LowerDmlToAst to take a new parameter. To make room for it without
changing the lowering API outside of introspection, this commit removes
the LowerDmlToAst struct and replaces it with a free-standing function
taking a context-like arguments (LowerParams).

The resulting code is more consistent with the patterns we use in the
parser database and validation, and arguably simpler, easier to follow.