feat(datamodel): reformat after rendering datamodel
In introspection, we use lowering, the AST renderer from schema-ast to
produce the final prisma schema string we return to the user. This is
one implementation of prisma schema formatting.

We also have the reformatter used by prisma-fmt. The two formattings
produced by these two components are different. We want to align on a
single formatting eventually: this is the purpose of this commit.
The only change in this commit is that we run introspected schemas
through the reformatter.

After this, we are free to stop caring about indentation and formatting
details in schema_ast::renderer, since these are concerns for the
reformatter.

Relevant issue: https://github.com/prisma/prisma/issues/12580