refactor(datamodel): delete unused PSL grammar rule
That rule was never used — there was code using it, which was never
reached. This commit deletes it instead of plugging it back in, because
it can be tackled more thoroughly and efficiently with a different
approach, namely by validating identifiers _in general_. We would
include validating that there are no leading digits, but also no leading
underscores, and only legal characters.