refactor(reformat): completely separate relation additions from AST reformatting
That way we can do pure AST reformatting if we want, and reuse the logic
for relation and relation scalar field additions in code actions. As a
side-effect, that makes the formatter idempotent.

There will be a follow-up PR to split the two functionalities in two
different modules, but I stopped here for readability in this PR.

closes https://github.com/prisma/prisma/issues/12726