fix(sql-schema-describer): always return foreign keys sorted
This caused introspection to not find foreign keys that were in the
SqlSchema (see
https://github.com/prisma/introspection-ci/commit/80965f82d27159a9b8d40818f0b6e6387f62789d) on postgres and mysql.

The error was not caught because we were not testing with large enough
schemas with constraint names repeating between tables. We import the
northwind schema for mysql and postgres in introrspection tests, since
it did exhibit these bugs.