refactor(datamodel): render datamodels without AST
This change was made possible by https://github.com/prisma/prisma-engines/pull/3049

This is a massive simplification of the implementation of rendering a
datamodel to a string, that has the additional advantage of decoupling
the AST from rendering concerns. We can now always assume AST nodes are
backed by a source document. This makes it possible to drop a large
amount of extraneous logic in the AST and dml.

The API surface of the schema_ast crate shrinks considerably as a result.