refactor(datamodel): shake off the dependency on thiserror in diagnostics
This pulls in the whole procedural macro machinery before we can even
get to compile schema_ast. Removing it should improve compile times
considerably.

From a logic perspective, ErrorKind was implemented as an std::Error
implementation a long time ago, but it is the wrong tool for parser
or validation errors, since they are not logic errors, they are an
expected output of the parser, that we just accumulate.