refactor(datamodel): absorb DatasourceProvider trait in Connector
Before this commit, we had two traits implemented for each datamodel
provider: Connector and DatasourceProvider. Connector was implemented in
separate crates for the builtin providers, but DatasourceProvider was in
core. With this commit, we only have Connector, which makes sure the
connector-specific logic is in only one place.

We now only have Connector. This unlocks further simplifications beyond
the scope of this PR.