Add tracing path from engine to client (#3022)
This adds useful tracing spans to the engine to observe the flow
of prisma queries. A tracer is implemented for the node-api that passes
the traces to the log callback function as json so that the prisma client
can process them and add them to the OTEL pipeline.

On the queryengine binary side, the traces are written to stdout via
a new tracer. This again allows the client to read the traces and
add the tracing spans to the OTEL pipeline.


Co-authored-by: daniel starns <danielstarns@hotmail.com>