feat: add a code-free test setup for introspection-engine-tests
SQL -> introspected prisma schema. We expect better compile times and
fewer diffs due to meaningless changes in test setup after this.

It also makes it easy to add additional checks (introspected schemas are
valid, etc.) without adding or changing calls in tests.

The first test is the reproduction of a regression in indexes
introspection.