me/ie/describer: make mssql use one database per test
This is necessary because we are going to work on multi-schema support,
and we will have to test that.

Command: hyperfine -w2 --show-output 'mold -run cargo test'

introspection-engine-tests, this branch, mssql 2017:

```
  Time (mean ± σ):     21.419 s ±  0.597 s    [User: 4.892 s, System: 0.329 s]
  Range (min … max):   20.315 s … 22.272 s    10 runs
```

On main, the fastest I managed the introspection tests to go, with 2
threads to avoid deadlocks, was this:

```
  Time (mean ± σ):      6.911 s ±  0.145 s    [User: 4.609 s, System: 0.241 s]
  Range (min … max):    6.689 s …  7.081 s    6 runs
```

So database-per-test is definitely slower than schema-per-test, but we
do not have a choice, because we will support and test multiple schemas.