refactor(schema-ast): make a few fields private
Background: we find it very convenient in other places in the codebase
(sql-schema-describer, mongodb-schema-describer,
sql-migration-connector, datamodel-core, parser-database) to have
lightweight, stable identifiers for precisely and conveniently keeping
references to specific elements in a Prisma schema or a database schema.

For the Prisma schema part, we are limited by the current shape of the
AST: we would like to refer to a specific argument inside a field
definition in an index attribute, for example, and with the current data
structure, it would be possible but very inefficient and verbose (see
the already gnarly `AttributeId` enum we have). Like we did in the rest
of datamodel, we need to progressively change the schema AST to a more
uniform representation that supports stable identifiers, and as a
consequence easy traversing (field would have a method to access its
parent, etc., like we already have in parser-database and
sql-schema-describer).

This PR: since this is a big undertaking, this PR starts with a simple
step: making the concrete shape of the AST private, and make consumer
code use methods to access specifics. This PR is just a start, but with
consistent work, we will get to a point where we can change the
underlying AST data structure without breaking downstream crates (mainly
parser-database and datamodel-core).