refactor(me/sqlite): remove one level of error transformation
...for better error messages

We now handle rusqlite errors directly instead of relying on quaint.
This generally gives us more flexibility and more context on the errors
we report, less noise. The idea is to progressively do this for the
other connectors as well. On SQLite, another benefit of this change is
that we can make some asynchronous code synchronous.

This commit is part of the general effort to remove unnecessary layers
of abstraction in the migration engine: we have a connector abstraction,
everything below that can and should be connector specific, there is
nothing to be gained from e.g. extra layers of error translation.