dml-renaming: start preparing datamodel tests to move to psl
The big difficulty is that the tests are not consistent in how they use
the datamodel API and how they are set up, so it is going to be a lot of
manual work. We have convenient test helpers in `common` that we can use
as a level of indirection, so this PR makes more tests use them.

There is still a long road ahead.

Eventually, we want to eliminate accidental complexity like variable
names and the choice of how you get and render datamodel errors from
datamodel tests. Something like the new introspection engine simple test
setup would make the most sense.