sql-schema-describer/pg: fix index query failure w/ intarray ext
The intarray extension defines its own `@>` operator for integer arrays,
and that conflicts with the builtin postgres `@>` operator — the query
fails with a call to disambiguate using the `OPERATOR(_, _)` construct,
which is unfamiliar to most people. This commit replaces that usage of
`@>` with a more readable call to `array_position()`.

closes https://github.com/prisma/prisma/issues/14389