sql-schema-describer(mssql): filter out INCLUDEd columns from indexes
On MSSQL, indexes can have an `INCLUDE` clause. Columns listed in that
clause are _not part of the key_, meaning they are not indexed, but they
are stored at the leaf. This is useful for queries that only care about
columns that are part of the key or INCLUDed: they will not need to
fetch the whole tuple from the main table, the query can be resolved
purely from the index. This commit changes sql-schema-describer to
ignore included columns: they are not part of the index and should not
show up in introspected schemas.

Before Prisma 4.1, we already erroneously introspected the included columns
as part of the index. Due to a data structure change in 4.1, we
continued wrongly introspecting them, but _at the end of a different
index_ (the previous one), potentially on another model, triggering the
error about "your friendly prisma developers" you can see in the issues
below.

See the official docs on included columns:
https://docs.microsoft.com/en-us/sql/t-sql/statements/create-index-transact-sql?view=sql-server-ver16#include-column---n--

closes https://github.com/prisma/prisma/issues/14636
closes https://github.com/prisma/prisma/issues/12615
closes https://github.com/prisma/prisma/issues/14438
closes https://github.com/prisma/prisma/issues/14403
closes https://github.com/prisma/prisma/issues/14511
closes https://github.com/prisma/prisma/issues/14611
closes https://github.com/prisma/prisma/issues/14647
closes https://github.com/prisma/prisma/issues/14462