psl: add direct access to fields to IndexWalker
We have too many layers of indirection, and are losing too much source
information in parser-database. The layers come from from needing
specific information in validations that needed to be resolved
_somewhere_, and that information was lost after parser-database, and
partly from not having a good AST data structure, that makes us lose
that information in the first place. This PR adds a method to dig
through the AST to the `fields: [...]` array inside an index. This will
become significantly easier when we replace the current SchemaAst
structure.