sql-schema-describer: remove barrel dependency
We continue moving to more database-specific, more declarative tests in
the introspection and migration engines. Barrel is not maintained and
causes more friction than it helps for more specific tests.

One possibly controversial move in this topic is that some tests were
deleted. The reasoning behind this is that what they test is tested ten
times over in more maintainable and specific ways in the introspection
test suite. Please voice any concern.