qe: shave another 30%+ off query-engine-tests compile times
We do this by generating a call to a regular function in the
connector_test macro, instead of generating the same async function
implementation once per test. The compiler has a lot less code to
compile. In query-tests-setup, we then type-erase the future for the
test body, meaning the compiler only needs to compile the test setup
function exactly once for the whole test suite, instead of once per
test.