qe: fix prisma 14696
See this internal doc for the explanation of the bug, its symptoms, its
causes and the suggested fix: https://www.notion.so/prismaio/Repair-Prisma-14696-relation-ambiguity-e881652fc48b48e69bcad63f23fb296a#44c9b285938c4c2db8cd2e386dca5df7

The TL;DR is that the PSL validations and the QE don't agree on the
scope of relation names (`@relation("this-is-the-name")`). This caused
the QE to pick the wrong relation to follow in cases where two relation
fields on the same model shared the same relation name without being in
the same relation. This lead mostly to panics, but also in some cases in
writes on the wrong table.

closes https://github.com/prisma/prisma/issues/14696