qe+postgres: exclude `contains` filters on uuid fields
Internal doc on the problem and suggested solution: https://www.notion.so/prismaio/UUID-filters-in-Client-90da479ef4014b20bb846cbb572fb47e

The story on how to fix this became slightly more complicated than in
the spec.

The problem is that the string filters are registered on a global
`StringFilters` GraphQL input object type. All string fields have a
`StringFilters` input type for filtering over them. There cannot be two
different `StringFilters`, because this is not possible in the first
place in a GraphQL schema, and besides that input object types are
cached/memoized by name in schema-builder. For the adapted UUID filters,
we have to create a `UuidFilters` input type. The `scalar_filter_name()`
method on Connector takes care of that. It can evolve towards a more
typed API over time.

closes https://github.com/prisma/prisma/issues/9007