refactor(psl): implement single definition for native types
The implementation of native types in PSL is spread between different
methods and data structures: the `native_types` crate,
`Connector::introspect_native_type()`,
`Connector::available_native_types()`, `Connector::parse_native_type()`.

These definitions are manually written and required to be in sync. This
is tedious and error prone.

Moreover, when native types are part of data structures not part of the
connector, they are serialized to a `serde_json::Value`, leaving open
the opportunity for leaking implementation details and breaking
abstraction boundaries.

In this commit, we move the native type definitions for each connector
in the corresponding module of `psl/builtin_connectors`. A macro is
defined that encapsulates all the aspects handled by the items
mentioned above that had to be kept in sync manually. The JSON
representation of native types is represented by opaque existential
types that can be downcast by the connector that instantiated them, in a
similar approach to what we use in other core <> connector API
boundaries.

**This is not the ideal API**. The API in this commit tries hard not to
depart too much from the current `NativeTypeInstance` type, to avoid an
even larger refactoring. There are two directions we should consider in
the not too distant future:

- Stop storing native types in native type instances. Use AST node IDs
  to refer to the corresponding attributes. `NativeTypeId` + a `NodeId`
  for arguments would be a type-erasure free alternative that would be
  lighter and more type-safe. What prevents this for now is the
  difficulty of identifying nodes precisely with the current AST data
  structure (it will become easier with upcoming refactorings).
- Using concrete generics to specialize data structures by connector,
  e.g. `ParserDatabase<T>`. With an associated type on `T`, we can make
  the type erasure of native types unnecessary.