ie: unify and enrich internal introspection params
This is an internal change that does not alter the public API of the
introspection engine.

We took only the dml of the previous schema previously, now we take the
ValidatedSchema as well. It has much more context in a much nicer API
for reintrospection. Taking a `ValidatedSchema` directly to construct an
IntrospectionContext instead of a variety of inputs also ensures that
the datasource, preview features, datamodel, schema source string and
validated schema all come from the same input string, that they are
consistent.