qe: fix create w/ @default(now())/@updatedAt in id
This is a regression introduced in
fcf3572335e1fd8e0cd0fff9322345dae114f071. `@default(now())` and
`@updatedAt` timestamps became more precise. This caused an issue with
create queries. Schematically, this is what we do:

- INSERT with the high precision timestamp
- SELECT back the record using its primary key columns (the id fields)
  - the SELECT has a WHERE clause that uses the high precision timestamp
    we used

The catch with this process is that the database has stored a _lower
precision_ timestamp in the column, so the WHERE does not match the
record that was just inserted, and we get the error from
https://github.com/prisma/prisma/issues/15581

This commit makes the timestamps for `@default(now())` and `@updatedAt`
values less precise (millisecond precision only), re-establishing the
equality comparisons. This approach was proposed and originally decided
against in
https://www.notion.so/prismaio/QE-now-changes-within-the-same-request-280b56d1075f43dea5c5dd82b755541b

Internal design doc: https://www.notion.so/prismaio/Single-create-with-default-now-or-updatedAt-in-id-fail-d6550337014c4e5ab81d4d362228aa14

closes https://github.com/prisma/prisma/issues/15581