psl: fix database name clash validation with multiSchema
The validation we have validates that no two models in the whole schema
have the same database name. Database name is defined as the contents of
the model's `@@map` attribute, or the model name if the model name does
not have an `@@map` attribute. It is the name we give to the
table/collection corresponding to the model in the database.

When the `multiSchema` preview feature is turned on, the tables in the
database can be in different schemas, so we have to adapt the
validation to allow multiple models to have the same database name, as
long as they are in a different schema. This is what this commit does.

closes https://github.com/prisma/prisma/issues/15009