refactor(introspection): introspect enums in the right order
Instead of reordering the enums after introspection in the
re-introspection phase, this commit gives the introspected enums the
right order from the start.

This is important because the enum index in the schema becomes a stable
identifier.

This is a test run before we do this for models, which we will want for
multi-schema introspection.