Migrations, multi-schema: pass namespaces around as needed (#3331)
This enables basic functionality of migrations for `MultiSchema`. There is one test added, which has a multi-schema file with three schemas. We test that once we run a migration, the follow-up will be idempotent (result in an empty migration).

It is almost certain there's other things that need to happen before we can generally say migrations work for multi-schema, but this felt like a good first step.

I left in a bunch of `TODO(MultiSchema)` because I am unsure if and what needs to be done there. I think those are fine to stay there, at least temporarily, until we think we're (mostly) done with multi-schema migrations, as they might be good "review" anchors for that.