Fix broken main test from multiple non-conflicting PR changes (#3348)
See this comment https://github.com/prisma/prisma-engines/pull/3347/files#r1007951554

And this slack message:

It doesn't happen very often but it has happened a few times in the past, and just now we have another case (https://github.com/prisma/prisma-engines/pull/3347/files#r1007951554): sometimes PRs aren't rebased, the tests are green, but once they are merged, they fail on main. In this case because one test made assumptions that worked with the older main the branch was based on, but not the new ones. All without merge conflict.I think you can enforce that branches are green and rebased/up-to-date with main before merging in GitHub. Should we do that? The obvious downside is that most of the time, it will just make PRs take longer to merge as you have to wait for CI to be green again after the rebase.